+++ packages/cashier/src/pages/payment-agent/payment-agent-list/payment-agent-list.jsx
@@ -9,7 +9,7 @@ import VerificationEmail from 'Components/verification-email';
 import PaymentAgentContainer from '../payment-agent-container';
 import PaymentAgentWithdrawalLocked from '../payment-agent-withdrawal-locked';
 import PaymentAgentDisclaimer from '../payment-agent-disclaimer';
-import { SignupSideNote as PaymentAgentSignupSideNote } from '../payment-agent-signup';
+import { PaymentAgentSignupSideNote } from '../payment-agent-signup/components/signup-side-note';
 import SideNote from 'Components/side-note';
 import './payment-agent-list.scss';
