+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/components/cancel-wizard-dialog/cancel-wizard-dialog.tsx
