+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/components/signup-side-note/index.ts
@@ -1,0 +1,3 @@
+import SignupSideNote from './signup-side-note';
+
+export { SignupSideNote as PaymentAgentSignupSideNote };
