+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/components/signup-side-note/signup-side-note-provider.tsx
