--- packages/cashier/src/pages/payment-agent/payment-agent-signup/signup-side-note/index.ts
@@ -1,3 +1,0 @@
-import SignupSideNote from './signup-side-note';
-
-export default SignupSideNote;
