--- packages/cashier/src/pages/payment-agent/payment-agent-signup/signup-wizard-steps/select-country-step/index.ts
@@ -1,3 +1,0 @@
-import SelectCountryStep from './select-country-step';
-
-export default SelectCountryStep;
