+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/signup-wizard.scss
