--- packages/cashier/src/pages/payment-agent/payment-agent-signup/signup-wizard/index.ts
@@ -1,3 +1,0 @@
-import SignupWizard from './signup-wizard';
-
-export default SignupWizard;
