+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/steps/country-of-issue/index.ts
@@ -1,0 +1,3 @@
+import CountryOfIssue from './country-of-issue';
+
+export default CountryOfIssue;
