+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/steps/selfie/index.ts
@@ -1,0 +1,3 @@
+import Selfie from './selfie';
+
+export default Selfie;
