+++ packages/cashier/src/pages/payment-agent/payment-agent-signup/steps/selfie/selfie.scss
